:recycle: refactor: Rename packages - inbound -> in - outbound -> out - rest -> web
