--- kok-api/src/main/java/com/kok/kokapi/config/UseCaseConfig.java
@@ -1,15 +1,0 @@
-package com.kok.kokapi.config;
-
-import com.kok.kokapi.room.application.service.RoomService;
-import com.kok.kokcore.application.port.outbound.RoomRepository;
-import com.kok.kokcore.usecase.CreateRoomUseCase;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class UseCaseConfig {
-    @Bean
-    public CreateRoomUseCase createRoomUseCase(RoomRepository roomRepository) {
-        return new RoomService(roomRepository);
-    }
-}
