+++ kok-api/src/main/java/com/kok/kokapi/config/usecase/UseCaseConfig.java
@@ -1,7 +1,7 @@
 package com.kok.kokapi.config.usecase;
 
 import com.kok.kokapi.room.application.service.RoomService;
-import com.kok.kokcore.application.port.outbound.RoomRepository;
+import com.kok.kokcore.application.port.out.RoomRepository;
 import com.kok.kokcore.usecase.CreateRoomUseCase;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
